Move weird serivces into the app folder, remove abstraction since we don't need them.
