+++ apps/web/src/app/accounts/login.component.ts
@@ -19,8 +19,8 @@ import { Policy } from "@bitwarden/common/models/domain/policy";
 import { ListResponse } from "@bitwarden/common/models/response/listResponse";
 import { PolicyResponse } from "@bitwarden/common/models/response/policyResponse";
 
-import { StateService } from "../../abstractions/state.service";
 import { RouterService } from "../services/router.service";
+import { StateService } from "../services/state.service";
 
 @Component({
   selector: "app-login",
