+++ apps/web/src/app/models/globalState.ts
