+++ apps/web/src/app/services/init.service.ts
@@ -17,7 +17,7 @@ import { ContainerService } from "@bitwarden/common/services/container.service";
 import { EventService as EventLoggingService } from "@bitwarden/common/services/event.service";
 import { VaultTimeoutService as VaultTimeoutService } from "@bitwarden/common/services/vaultTimeout.service";
 
-import { I18nService as I18nService } from "../../services/i18n.service";
+import { I18nService } from "./i18n.service";
 
 @Injectable()
 export class InitService {
