+++ apps/web/src/app/services/passwordReprompt.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@angular/core";
 
 import { PasswordRepromptService as BasePasswordRepromptService } from "@bitwarden/angular/services/passwordReprompt.service";
 
-import { PasswordRepromptComponent } from "../app/components/password-reprompt.component";
+import { PasswordRepromptComponent } from "../components/password-reprompt.component";
 
 @Injectable()
 export class PasswordRepromptService extends BasePasswordRepromptService {
