+++ apps/web/src/app/services/webPlatformUtils.service.ts
