+++ apps/web/src/app/settings/settings.component.ts
@@ -6,7 +6,7 @@ import { OrganizationService } from "@bitwarden/common/abstractions/organization
 import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUtils.service";
 import { TokenService } from "@bitwarden/common/abstractions/token.service";
 
-import { StateService } from "../../abstractions/state.service";
+import { StateService } from "../services/state.service";
 
 const BroadcasterSubscriptionId = "SettingsComponent";
